+++ src/pages/Funding/FundingCreate/FundingCreateStyles.js
@@ -3,7 +3,7 @@ import styled from "styled-components";
 // 전체 컨테이너
 export const MainContainer = styled.div`
   display: flex;
-  justify-content: center; 
+  justify-content: center;
   max-width: 1200px;
   min-height: 100vh;
   margin: 0 auto;
@@ -130,28 +130,28 @@ export const SponserComment = styled.div`
 `;
 
 export const InputTag = styled.input`
-    width: ${(props) => props.w};
-    height: ${(props) => props.h};
-    background-color: #eae7de;
-    border-radius: 4px;
-    border: none;
-    margin-left: ${(props) => props.ml};
-    margin-bottom: ${(props) => props.mb};
-    padding-left: ${(props) => props.pl};
-    padding-top: ${(props) => props.pt};
-    padding-bottom: ${(props) => props.pb};
-    font-weight: 500;
-    font-size: 12px;
-    justify-content: start;
-    align-items: start;
-    &:hover {
+  width: ${(props) => props.w};
+  height: ${(props) => props.h};
+  background-color: #eae7de;
+  border-radius: 4px;
+  border: none;
+  margin-left: ${(props) => props.ml};
+  margin-bottom: ${(props) => props.mb};
+  padding-left: ${(props) => props.pl};
+  padding-top: ${(props) => props.pt};
+  padding-bottom: ${(props) => props.pb};
+  font-weight: 500;
+  font-size: 12px;
+  justify-content: start;
+  align-items: start;
+  &:hover {
     cursor: pointer;
   }
 `;
 export const RadioInput = styled.input`
-    margin-bottom: ${(props) => props.mb};
-    accent-color: black;
-`
+  margin-bottom: ${(props) => props.mb};
+  accent-color: black;
+`;
 
 export const FundingNewline = styled.div`
   width: 100%;
