+++ src/pages/Funding/FundingPay/CheckBox/CheckBoxStyles.js
